#365 - Use role info from LTI
Instead of using the role stored in the DB, use the role provided in the LTI launch parameters.  The LTI params are available in the session after the initial launch request.  The role in the DB doesn't seem to be updated after it's stored there by the initial user record creation.  In local development, the change to `djangolti/backends.py` didn't seem to have any affect.  It may work on the OpenShift environment.