#365 - Removed role assignment
Removed role assignment.  Setting the roles shouldn't be a problem, but it's unnecessary if most of the code is using the LTI role from the session.  However, some part of the code is apparently still referring to this role, because when a user has multiple roles, the app gives permission error messages.