#365 - Remove ineffective query param
Not actually related to the role issue.  However, it seems like such a small change to open a new issue, branch, etc. for it.  I was reminded by an error email that included the query with a "per_page=500" parameter.  Instructure allows a maximum page size of 100 results from Canvas API queries.  I changed this section of the code and commented it to remove confusion.