#376 - Return data for manual assign reviews
Two parts needed to be updated to show manually assigned reviews.  The reviews needed to be added to the list of reviews for a rubric.  The views of the student needed to show data about each of the reviews assigned.  The Django ORM is awkward to work with at times, but it's made a little more difficult by the perspective of the existing code.  It was written based on the assumption that only students with submissions would do reviews.  The manual assignment without submission is a new feature that breaks that assumption.  Maybe the code should have been written (or now rewritten) to look up assigned reviews for a rubric regardless of submissions.