#384 - new HTTP result code
Nagios uses outdated curl/wget to check on services.  When the service returns a 500 HTTP result code, the tools are unable to return the content of the response.  Therefore, any diagnostic information in the content is lost.  Until Nagios is upgraded or gets new tools, change Watchman configuration to use a 203 HTTP result code instead.  The tools seem to work well with 2xx codes, so the variation from a 200 HTTP result code may be a helpful indicator that all is not quite correct, especially when used in conjunction with the contents of the response.