#256 - fix textarea
This textarea wasn't showing existing content.  Vue doesn't allow the value enclosed in the `textarea` elements, but browsers don't support the `value` property.  Added JS code to copy the content from the `value` property to where it belongs.

The use of the `label` element here didn't seem right and the way Vue/MDL manages it would interfere with proper operation.  Switched to using standard HTML `placeholder` attribute instead.   The MDL `textfield` style on the textarea also interfered with the display of existing content, so removed that, too.

In the past, instructors commented that the single underline style of the textarea wasn't a good enough visual indicator that it's where the input should be written.  Removed that style.