#256 - fix annoying eslint rules
It's annoying that eslint complains about trailing commas on structures, so I enabled some.  I also strongly disagree that it's an error to use a space after `if`, `for`, etc.  (They are keywords, NOT functions!)  I think they should be required, but I wasn't prepared to go through all the code and correct them all.  So, for now, I just changed the severity from `error` to `warn`.