#451 - corrected `submitted` logic
In Canvas, some students who did not make a submission may have received a grade (probably a zero) on the prompt assignment.  That grade changes their `workflow_state` in Canvas API responses from `unsubmitted` to `graded`.  MPR's logic was not compatible with this.  We're not sure whether this logic was always bad or Canvas changed the values returned by its API.