#457 - ignore submissions from unknown students
When the Canvas "Test Student" feature is used to make submissions to prompt assignments, those submissions are available to the review distribution process, however, the test student is not part of the class list.  When MPR tries to apply the submission, but cannot find the student, it causes an error.  These steps to remove submissions from unknown students will prevent that error.  Also added useful comments and log statements.