#378 - apply migrations at launch
To make deploying updates with migrations easier, added code to run `manage.py migrate` when the API container is launched.  This also required code to wait until the DB is ready for use, which requires jq to read the DB host and port from the configuration file.  This may need some adjustment; it has not yet been tested in OpenShift environment.