correct name field lengths
The migration had been created based on an old DB structure, so the lengths of `first_name` and `last_name` in `auth_user` have been set to 30 instead of 150.  When applied to production data, this migration fails because some values in those columns contain more than 30 characters.  Changing these back to 150.